mexicoder/requirements.txt
Django==1.7.1
Markdown==2.5.1
South==1.0.1
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.18
celery==3.1.16
celery-with-redis==3.0
dj-database-url==0.3.0
dj-static==0.0.6
django-celery==3.1.16
django-filter==0.8
django-toolbelt==0.0.1
djangorestframework==2.4.4
gunicorn==19.1.1
kombu==3.0.23
psycopg2==2.5.4
pytz==2014.9
redis==2.10.3
static3==0.5.1
wsgiref==0.1.2
